fix(redis): 统一 Redis 过期时间设置类型
- 将 EmailTrackingService 中所有 Redis 设置的过期时间参数类型由秒数改为 Duration 类型
- 确保邮件跟踪数据在 Redis 中的 TTL 设置规范且一致
- 修正 JobRepository 中分页信息的类型，将页码和页面大小从 int 转换为 long 类型
- 提升代码类型安全性，避免类型不匹配导致的问题

+++ ydsz-cronjob/ydsz-cronjob-domain/src/main/java/com/njydsz/cronjob/domain/repository/JobRepository.java
@@ -362,8 +362,8 @@ public interface JobRepository {
       response.setMsg("操作成功");
       response.setData(records);
       response.setTotal(total);
-      response.setPageNum(pageNum);
-      response.setPageSize(pageSize);
+      response.setPageNum((long) pageNum);
+      response.setPageSize((long) pageSize);
       return response;
     }
   }
